# services/service-router/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Service Router API",
    "description": "This API is used to register services and to expose information about them.",
    "license": {
      "name": "TODO",
      "url": "http://todo.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:6000",
      "variables": {}
    }
  ],
  "paths": {
    "/servicerouter/v1/register": {
      "post": {
        "summary": "registerService",
        "description": "Register a new service",
        "operationId": "registerService",
        "parameters": [],
        "requestBody": {
          "description": "Service object to register",
          "content": {
            "application/json": {
              "schema": {
                "description": "Service object to register",
                "$ref": "#/components/schemas/Service"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Register a Service",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Service object to register",
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Bad Request Error",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "deprecated": false,
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Service": {
        "title": "Service",
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "definitionPath": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "latency": {
            "type": "integer"
          },
          "isHealthy": {
            "type": "boolean"
          } 
        }
      },"Error": {
        "title": "Error",
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  } 
}